client should attempt graceful quit
We were only calling process.Kill before which kills the process without
allowing it to perform any cleanup (defers are never called). This was
leaving around a lot of temporary files over time.

We now add a graceful quit method to the control stream that we attempt
to call first. If that fails, we still force quit, but at least this
gives the process a chance to gracefully exit.