client.Kill should check for addr before closing Client
This fixes an issue that Terraform users were seeing wit periodic hangs.
The unit test without this fix reproduces this hang.

This situation was caused when a plugin fails to start properly (invalid
RPC protocol version for example). In this case, `Start` sets `process`
since the binary started, but it never sets `address` because the plugin
is incompatible and never gets to serve an address.

When calling `Kill` in this situation, it wouldn't no-op since process
is set. Next, it'd call `Client` to get the client to close the
connection, but since `address` is nil Client would start the plugin
binary _again_. This would put the entire Client into a bad state and
would case Kill to fail and hang.

There are more bugs here for sure (not new ones introduced, just others
I found) but I want to fix this issue in isolation first.