value in log kv pairs can be any type
Fixes #38

hclog can properly handle interface{} types. Therefore, don't try to
turn the type into a string and just serialize the interface{}.

The serialization is done via JSON on stderr so this will effectively
and safely support any Go type witout panic. On the other side, hclog
will do a `fmt` with `%#v` for any type it can't natively support
outputting. This makes it safe to send directly to hclog.

While I was doing this work, I modified the test to also run against
grpc, so we can verify that grpc plugin logging works as well. This
should've been the case (and was) since the logging code is not specific
to protocols, but its good to gain coverage.