set the default log level to hclog.Trace
The legacy behavior for go-plugin was to log everything to the
default log output. The default hclog.Logger has a log level of Info, so
clients that have not updated to provide a custom Logger no longer will
see any output from the plugins.

This updates the default Logger to Trace, so that all output is still
seen by clients.