stderr may be json, but not key-value pairs.
The isJSON checked if the output was valid json of any type, but the
logger only unmarshaled into `map[string]interface{}`. If the line
didn't umarshal correctly, the error was only logged, but the actual
line was lost and the loop continued into a nil pointer dereference.

Always log the raw line if it can't unmarshal into a map.