Make the tests pass the race detector (#63)
These changes are relatively obvious; the one that might cause a little
heartburn is declaring a package-level variable for test stdout/err, but
given the control flow of the test functions it's either that or modify
the test function APIs. I'm happy to do the latter if it's deemed to not
be a big deal, since they're test functions.