add prtocol version negotiation to plugins
This add protocol version negotiation to the go-plugin package.
Version negotiation is backwards compatible with the current handshake
configuration, and allows both client and server to have multiple sets
of plugins, and choose the latest compatible version to use.

The set of versions implemented by the client is communicated with the
server via the `PLUGIN_PROTOCOL_VERSION` env variable, and the chosen
pluginversion is returned to the client using the existing wire
protocol.