make sure Kill is idempotent
Client.Close will issue warnings if called after the client is killed.
Kill is safe to call multiple times, so make sure we remove the process
after the first call so that subsequent calls are a noop.

Note that the plugin process will never exit gracefully for now.

Defer context cancellation so that it is easier to visually determine
that it is called at the correct time.