fix shutdown notification in grpc server
The Serve method was waiting on a channel that was never closes. The
Serve method is what should be closing the channel to notify when the
server is done.

The stdrr read loop only checks for EOF, so if there's another type of
error, it ends up in a busy loop.