Further shutdown and test fixes
Simplify the goroutine management by removing multiple channels and
explicitly waiting for _all_ client goroutines to return on Kill().

Make sure Context cancellation funcs are called in all paths.

Process termination it detected using the existing cancellation context.
This makes detection of early termination and termination during Kill
use the same mechanism, rather than the latter relying to the
side-effect of the logging channel being closed.

Add a flag to mark when a process was forcefully killed for tests. (this
addition makes the pre-refactor grpc tests fail, to validate some of the
PR changes)