add the grpc controller for server process control
Unlike the netRPC server, the gRPC server cannot terminate when the
client closes its connection. Since we want to use the server as a
plugin with control over its lifecycle, there needs to be another
channel of communication to notify the server when to shutdown.

This adds a small grpc service, currently with only a Shutdown method,
which calls Stop on the GRPCServer. This will allow the server to
shutdown normally without the client timing out and calling Kill on the
process.