gRPC stdout/stderr syncing support
This is a feature that has existed in go-plugin for a few years by
setting SyncStdout, SyncStderr on ClientConfig, but this feature only
worked with net/rpc plugins previously.

This feature will sync the os.Stdout/Stderr output from the plugin to
the specified writers. If no writers are specified, it drops the data.
Note that the os.Stdout/Stderr that is synced is the overridden writers
that you can see set in server.go around line 250 at the time of this
commit.

The use case for this feature is that a plugin that subprocesses for
example can connect stdout/stderr and the output of the subprocess will
reach the target synced stream (which may just be the host process
stdout/stderr).