Add ServeOption to disable syncing stdout.
By default, os.Stdout and os.Stderr are overwritten with handlers that
pass their output over gRPC to the client. When the server is running as
the main process, or running in the same process as the client, this has
unfortunate side-effects.

In this case, running a server in the TestMain process of `go test`
swallows some test output, include coverage information, which isn't
ideal.

This adds an option to disable this behavior when starting a server.