Add a test mode
This is a special configuration on the plugin server (the plugin side)
that enables running the plugin in-process and eases many of the
behaviors around that.

The desire to do this is for more easily using `go test` (and similar)
with plugins. This mechanism allows the plugin serving to start up
in-process, to grab a ReattachConfig, and to be able to send that
ReattachConfig to some other plugin client for connection.

Behaviorally, this prevents plugin serving from exiting the process with
os.Exit, prevents taking over os.Stdout in such a hostile way, prevents
connection info going to os.Stdout which floods the test logs, etc.