Add more helpful message on plugin setup failure
The current error message shown when a plugin does not respond correctly
can be obtuse to the end user of, for example, Vault.

Here, we add potential reasons why the plugin did not respond correctly,
and some additional debugging information (provided as a best effort)
including the CPU architecture that the plugin was compiled for,
the current CPU architecture, and the permissions of the plugin.
Hopefully this will help users diagnose why their plugin is not loading.

We also added a `testdata/` directory that is optionally populated
with executables in various formats to help test the additional
debugging information. The binaries are over 1 MB each though, so
are not checked in, and the test will be skipped if they have not been
compiled.