remove logging for missing `SecureConfig`
In #207 we introduced a warning if the `SecureConfig` is unset. For Nomad and
other applications that have "internal" go-plugin consumers where the
application runs itself as a plugin, this causes spurious warn-level logs. For
Nomad in particular this means every task driver and logmon invocation emits the
log, which is our primary operation.

Setting a `SecureConfig` isn't required, and if a given go-plugin consumer
requires it to be, this seems more like a programming error than a runtime
error. Remove the log line.