Fix golang grpc example's golang generation instructions
This was done using the latest release of protoc as of this authoring,
3.15.8. Whatever very old version of protoc was used before was not
stated, so in order to ensure that the example instructions still
worked, we wanted to use the latest available now. Updating the grpc and
protobuf dependencies was required, as well as adding the `go_package`
option to the .proto file.

This fixes #249.

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>
Co-authored-by: czavieh <czavieh@users.noreply.github.com>