Also fix bidirectional example in the same way
It's very similar to the gRPC example, and uses the same generation
process, so it needed to be updated in the same way. This wasn't on our
radar because we weren't referencing it, but after noticing that the
"negotiated" example was using the grpc example's code, we decided to
ensure all four examples were working as documented.

Of course, we also updated the README, and made sure to include a
.gitignore since this example was missing one.

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>