gRPC server supports `PanicHandler`
Supports `PanicHandler` for the gRPC server to allow users to catch and log the panic at the client side.

Currently, when the plugin panics, the panic is captured only by the stderr rpc channel, which is then logged by the client side with "DEBUG" log level. In most cases, the panic information should be of the "ERROR" log level by the definition of log level. This `PanicHandler` can help with that. In some other cases, this hook allows users to do other handling when the provider crashes.