Skip some logging work when log message would be ignored
In callers like HashiCorp Terraform, the plugin logger is completely
disabled (hclog.Off) by default, but previously go-plugin would nonetheless
waste time trying to parse log lines to decide what log level to use when
calling into the logger, creating memory allocation garbage and spending
time in the JSON parser only to immediately discard the results.

The hclog.Logger API exposes information about the log level so that
callers can skip doing expensive preparation work unless it will actually
be used. We'll now use that to totally skip all of the log preparation
when the logger is completely disabled, and to skip _some_ of the work
if the log line's level is not sufficient to match the logger's level.