Complete the implementation of moving panics to the Error log level (#353)
* cleanup panic logging flag

Don't shadow the builtin `panic` symbol.
Remove the redundant `false` assignments.
Flatten the if block since we are already using a switch.

* add Panic to grpc test service

* update impls with new Panic method

* test that server panics go to ERROR

* CHANGELOG

* fix test plugin mapping

* use Debug for logger test to avoid timed output

* Use buf to generate test proto (#354)

It wasn't clear which versions of the protobuf tooling needed to be used
to generate the test code, and updating them required some additional
changes. Since the `buf` tool was adopted to generate the example code,
use the same tool to make generating the test code easier too.